+++ src/ClosureInfo.php
@@ -99,15 +99,36 @@ final class ClosureInfo
         return ($this->flags & self::FLAG_HAS_SCOPE) === self::FLAG_HAS_SCOPE;
     }
 
+    public function getClosure(?array &$vars = null, ?object $thisObj = null, ?string $scope = null): Closure
+    {
+        return $this->getFactory($thisObj, $scope)($vars);
+    }
+
+    /**
+     * @internal
+     */
     public function getFactory(?object $thisObj, ?string $scope = null): Closure
     {
         $factory = ($this->factory ??= ClosureStream::factory($this));
 
-        if ($thisObj) {
-            return $factory->bindTo($thisObj, $scope ?? 'static');
+        if ($thisObj && $this->isStatic()) {
+            // closure is static, we cannot bind
+            if (!$scope) {
+                // we can extract scope
+                $scope = get_class($thisObj);
+            }
+            // remove this
+            $thisObj = null;
         }
 
-        if ($scope) {
+        if ($thisObj) {
+            if (ClassInfo::isInternal($thisObj)) {
+                return $factory->bindTo($thisObj);
+            }
+            return $factory->bindTo($thisObj, $thisObj);
+        }
+
+        if ($scope && $scope !== "static" && $this->hasScope() && !ClassInfo::isInternal($scope)) {
             return $factory->bindTo(null, $scope);
         }
 
@@ -147,7 +168,7 @@ return {$this->body};
     {
         $data = ['key' => $this->key()];
         if ($this->header) {
-            $data['imports'] = $this->header;
+            $data['header'] = $this->header;
         }
         $data['body'] = $this->body;
         if ($this->use) {
@@ -162,7 +183,7 @@ return {$this->body};
     public function __unserialize(array $data): void
     {
         $this->key = $data['key'] ?? null;
-        $this->header = $data['imports'] ?? '';
+        $this->header = $data['header'] ?? $data['imports'] ?? '';
         $this->body = $data['body'];
         $this->use = $data['use'] ?? null;
         $this->flags = $data['flags'] ?? 0;